fortress-vault-backend/requirements.txt
# Django Core
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Authentication
djangorestframework-simplejwt==5.3.1
djoser==2.2.2

# Database
psycopg2-binary==2.9.9

# Redis & Caching
redis==5.0.1
django-redis==5.4.0

# Celery
celery==5.3.6
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Channels (WebSockets)
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Utilities
python-decouple==3.8
Pillow==10.2.0
python-dateutil==2.8.2
requests==2.32.3

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Development
pytest==7.4.4
pytest-django==4.7.0
pytest-asyncio==0.23.3
factory-boy==3.3.0
faker==22.0.0
